+++ tests/testthat/test-read-oneshot-eav.R
@@ -5,7 +5,9 @@ update_expectation  <- FALSE
 
 test_that("smoke test", {
   testthat::skip_on_cran()
-  returned_object <- REDCapR:::redcap_read_oneshot_eav(redcap_uri=credential$redcap_uri, token=credential$token)
+  expect_message(
+     returned_object <- REDCapR:::redcap_read_oneshot_eav(redcap_uri=credential$redcap_uri, token=credential$token)
+  )
 })
 test_that("default", {
   testthat::skip_on_cran()