+++ NEWS.md
@@ -6,6 +6,9 @@ Upcoming Versions
 * New `redcap_delete()` function.  It deletes a vector of records.  (Thanks @joundso, #372, #373)
 * New `redcap_arm_export()` function.  It retrieves a list of REDCap project arms.  (#375)
 * `redcap_read()` and `redcap_read_oneshot()` accept a new `locale` parameter that specifies date, time, and number formats, like using a comma as the decimal separator.  It is a [`readr::locale`](https://readr.tidyverse.org/reference/locale.html) object.  (#377, suggested by @joundso)
+* New `redcap_instruments()` function exports a list of the data collection instruments for a project.  (#381, @vcastro)
+* New `redcap_event_instruments()` function exports the instrument-event mappings for a project (i.e., how the data collection instruments are designated for certain events in a longitudinal project)..  (#381, @vcastro)
+
 
 ### Minor Enhancements